+++ app/params/app_settings.go
@@ -5,6 +5,7 @@ import (
 
 	sdkmath "cosmossdk.io/math"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 )
 
 const (
@@ -12,10 +13,14 @@ const (
 	Name = "umee"
 
 	// BondDenom defines the native staking token denomination.
-	BondDenom = "uumee"
+	// NOTE: it is used by IBC, and must not change to avoid token migration in all IBC chains.
+	BondDenom      = "uumee"
+	BaseExtraDenom = "uux"
 
 	// DisplayDenom defines the name, symbol, and display value of the umee token.
-	DisplayDenom = "UMEE"
+	DisplayDenom = "UX"
+	// Old display name. We renamed UMEE to UX.
+	LegacyDisplayDenom = "UMEE"
 
 	// DefaultGasLimit - set to the same value as cosmos-sdk flags.DefaultGasLimit
 	// this value is currently only used in tests.
@@ -36,5 +41,34 @@ func init() {
 
 	if AccountAddressPrefix != Name {
 		log.Fatal("AccountAddresPrefix must equal Name")
+	}
+
+	sdk.SetCoinDenomRegex(func() string {
+		// allow "ux" token. Modify the default regexp to allow 2 character long denoms.
+		return `[a-zA-Z][a-zA-Z0-9/:._-]{1,127}`
+	})
+}
+
+// UmeeTokenMetadata creates bank Metadata for the UX token
+func UmeeTokenMetadata() banktypes.Metadata {
+	return banktypes.Metadata{
+		Description: "The native staking token of the Umee network.",
+		Base:        BondDenom, // NOTE: must not change
+		Name:        DisplayDenom,
+		Display:     DisplayDenom,
+		Symbol:      LegacyDisplayDenom,
+		DenomUnits: []*banktypes.DenomUnit{
+			{
+				Denom:    BondDenom,
+				Exponent: 0,
+				Aliases: []string{
+					"microumee", BaseExtraDenom,
+				},
+			}, {
+				Denom:    DisplayDenom,
+				Exponent: 6,
+				Aliases:  []string{LegacyDisplayDenom},
+			},
+		},
 	}
 }